# api/swagger.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Matrx Renderer API",
        "description": "HTTP API for managing and rendering Pixlet applications for Matrx devices",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8080",
            "description": "Local development server"
        }
    ],
    "paths": {
...
        "/apps/refresh": {
            "post": {
                "summary": "Refresh app registry",
                "description": "Reloads the app registry from the filesystem",
                "operationId": "refreshApps",
                "responses": {
                    "200": {
                        "description": "Apps refreshed successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/RefreshAppsResponse"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Failed to refresh apps"
                    }
                }
            }
        },
        "/apps/{id}": {
            "parameters": [
                {
                    "name": "id",
                    "in": "path",
                    "required": true,
                    "description": "App identifier",
                    "schema": {
                        "type": "string"
                    }
                }
            ],
            "get": {
                "summary": "Get app details",
                "description": "Returns detailed information about a specific app",
                "operationId": "getApp",
                "responses": {
                    "200": {
                        "description": "App details",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/AppManifest"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "App not found"
                    }
                }
            }
        },
        "/apps/{id}/schema": {
            "parameters": [
                {
                    "name": "id",
                    "in": "path",
                    "required": true,
                    "description": "App identifier",
                    "schema": {
                        "type": "string"
                    }
                }
            ],
            "get": {
                "summary": "Get app schema",
                "description": "Returns the schema definition for a specific app",
                "operationId": "getAppSchema",
                "responses": {
                    "200": {
                        "description": "App schema",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/AppSchema"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "App not found"
                    },
                    "500": {
                        "description": "Failed to load schema"
                    }
                }
            },
            "post": {
                "summary": "Validate schema configuration",
                "description": "Validates a configuration object against the app's schema",
                "operationId": "validateSchema",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/AppConfig"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Validation result",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ValidateSchemaResponse"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid JSON body"
                    },
                    "404": {
                        "description": "App not found"
                    },
                    "500": {
                        "description": "Failed to load schema"
                    }
                }
            }
        },
        "/apps/{id}/render": {
            "parameters": [
                {
                    "name": "id",
                    "in": "path",
                    "required": true,
                    "description": "App identifier",
                    "schema": {
                        "type": "string"
                    }
                }
            ],
            "post": {
                "summary": "Render app",
                "description": "Renders an app configuration and returns the base64-encoded WebP payload.",
                "operationId": "renderApp",
                "parameters": [
                    {
                        "name": "width",
                        "in": "query",
                        "required": false,
                        "description": "Device width in pixels (default 64)",
                        "schema": {
                            "type": "integer",
                            "format": "int32"
                        }
                    },
                    {
                        "name": "height",
                        "in": "query",
                        "required": false,
                        "description": "Device height in pixels (default 32)",
                        "schema": {
                            "type": "integer",
                            "format": "int32"
                        }
                    },
                    {
                        "name": "device_id",
                        "in": "query",
                        "required": false,
                        "description": "Optional device identifier used for logging",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/AppConfig"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Render result",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/RenderResponse"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid request"
                    },
                    "404": {
                        "description": "App not found"
                    },
                    "422": {
                        "description": "Validation failed",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ValidateSchemaResponse"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Failed to render app"
                    }
                }
            }
        },
...
        "/apps/{id}/call_handler": {
            "parameters": [
                {
                    "name": "id",
                    "in": "path",
                    "required": true,
                    "description": "App identifier",
                    "schema": {
                        "type": "string"
                    }
                }
            ],
            "post": {
                "summary": "Call schema handler",
                "description": "Calls a schema handler for dynamic field resolution",
                "operationId": "callSchemaHandler",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/CallSchemaHandlerRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Handler result",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/CallSchemaHandlerResponse"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid request"
                    },
                    "404": {
                        "description": "App or handler not found"
                    },
                    "422": {
                        "description": "Handler parameter validation failed (e.g. OAuth2 missing client_id, code_verifier, or client_secret)",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ValidateSchemaResponse"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Failed to invoke schema handler"
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "AppManifest": {
                "type": "object",
                "properties": {
                    "id": {
                        "type": "string",
                        "description": "Unique app identifier"
                    },
                    "name": {
                        "type": "string",
                        "description": "Display name"
                    },
                    "summary": {
                        "type": "string",
                        "description": "Short summary"
                    },
                    "description": {
                        "type": "string",
                        "description": "Detailed description"
                    },
                    "author": {
                        "type": "string",
                        "description": "App author"
                    },
                    "fileName": {
                        "type": "string",
                        "description": "Primary .star file"
                    },
                    "packageName": {
                        "type": "string",
                        "description": "Package name used internally"
                    },
                    "directoryPath": {
                        "type": "string",
                        "description": "Absolute path to the app directory"
                    },
                    "starFilePath": {
                        "type": "string",
                        "description": "Absolute path to the entry .star file"
                    }
                },
                "required": [
                    "id",
                    "name",
                    "summary",
                    "description",
                    "author",
                    "fileName",
                    "packageName"
                ]
            },
            "AppSchema": {
                "type": "object",
                "properties": {
                    "version": {
                        "type": "string",
                        "description": "Schema version"
                    },
                    "schema": {
                        "type": "array",
                        "description": "Configurable fields",
                        "items": {
                            "$ref": "#/components/schemas/AppSchemaField"
                        }
                    },
                    "notifications": {
                        "type": "array",
                        "description": "Supported notifications",
                        "items": {
                            "$ref": "#/components/schemas/AppSchemaNotificationField"
                        }
                    }
                },
                "required": [
                    "version",
                    "schema"
                ]
            },
...
            "AppConfig": {
                "type": "object",
                "description": "Arbitrary configuration object keyed by schema field IDs",
                "additionalProperties": true
            },
...
            "ValidateSchemaResponse": {
                "type": "object",
                "properties": {
                    "valid": {
                        "type": "boolean"
                    },
                    "errors": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/ValidationError"
                        }
                    },
                    "normalized_config": {
                        "$ref": "#/components/schemas/AppConfig"
                    }
                },
                "required": [
                    "valid",
                    "normalized_config"
                ]
            },
...
            "RenderResponse": {
                "type": "object",
                "properties": {
                    "result": {
                        "$ref": "#/components/schemas/RenderResult"
                    },
                    "normalized_config": {
                        "$ref": "#/components/schemas/AppConfig"
                    }
                },
                "required": [
                    "result",
                    "normalized_config"
                ]
            },
...
            "CallSchemaHandlerResponse": {
                "type": "object",
                "properties": {
                    "result": {
                        "type": "string",
                        "description": "Raw handler response"
                    }
                },
                "required": [
                    "result"
                ]
            },
            "RefreshAppsResponse": {
                "type": "object",
                "properties": {
                    "status": {
                        "type": "string",
                        "example": "success"
                    },
                    "message": {
                        "type": "string",
                        "example": "App registry refreshed successfully"
                    },
                    "app_count": {
                        "type": "integer",
                        "example": 42
                    }
                },
                "required": [
                    "status",
                    "message",
                    "app_count"
                ]
            },
...
        }
    }
}